fixes from code review

+++ internal/query/session_core.go
@@ -260,6 +260,10 @@ func (core *sessionCore) deleteSession(ctx context.Context) (finalErr error) {
 		return xerrors.WithStackTrace(err)
 	}
 
+	if core.id == "" {
+		return nil
+	}
+
 	_, err := core.Client.DeleteSession(ctx,
 		&Ydb_Query.DeleteSessionRequest{
 			SessionId: core.id,
